+++ cmd/networkcmd/status.go
@@ -42,7 +42,7 @@ func networkStatus(*cobra.Command, []string) error {
 	if err != nil {
 		return err
 	}
-	pChainBootstrapped, blockchainsBootstrapped, err := localnet.LocalNetworkHealth(app, ux.Logger.PrintToUser)
+	pChainBootstrapped, blockchainsBootstrapped, err := localnet.LocalNetworkHealth(app)
 	if err != nil {
 		return err
 	}
