+++ pkg/node/local.go
@@ -30,114 +30,6 @@ import (
 	"github.com/ava-labs/avalanchego/vms/platformvm/signer"
 )
 
-func TrackSubnetWithLocalMachine(
-	app *application.Avalanche,
-	clusterName,
-	blockchainName string,
-	avalancheGoBinaryPath string,
-	avaGoVersionSetting AvalancheGoVersionSettings,
-) error {
-	return localnet.LocalClusterTrackSubnet(app, blockchainName, clusterName)
-
-	if !localnet.LocalClusterExists(app, clusterName) {
-		return fmt.Errorf("local node %q is not found", clusterName)
-	}
-	// setup avalanchego
-	var err error
-	avalancheGoBinaryPath, err = setupAvalancheGo(
-		app,
-		avalancheGoBinaryPath,
-		avaGoVersionSetting,
-		ux.NoPrint,
-	)
-	if err != nil {
-		return err
-	}
-	return fmt.Errorf("PEPE")
-	sc, err := app.LoadSidecar(blockchainName)
-	if err != nil {
-		return err
-	}
-	clustersConfig, err := app.LoadClustersConfig()
-	if err != nil {
-		return err
-	}
-	clusterConfig := clustersConfig.Clusters[clusterName]
-	network := models.GetNetworkFromCluster(clusterConfig)
-	if network.ClusterName != "" {
-		network = models.ConvertClusterToNetwork(network)
-	}
-	if sc.Networks[network.Name()].BlockchainID == ids.Empty {
-		return fmt.Errorf("blockchain %s has not been deployed to %s", blockchainName, network.Name())
-	}
-	subnetID := sc.Networks[network.Name()].SubnetID
-	blockchainID := sc.Networks[network.Name()].BlockchainID
-	vmID, err := utils.VMID(blockchainName)
-	if err != nil {
-		return fmt.Errorf("failed to create VM ID from %s: %w", blockchainName, err)
-	}
-	var vmBin string
-	switch sc.VM {
-	case models.SubnetEvm:
-		_, vmBin, err = binutils.SetupSubnetEVM(app, sc.VMVersion)
-		if err != nil {
-			return fmt.Errorf("failed to install subnet-evm: %w", err)
-		}
-	case models.CustomVM:
-		vmBin = binutils.SetupCustomBin(app, blockchainName)
-	default:
-		return fmt.Errorf("unknown vm: %s", sc.VM)
-	}
-	rootDir := app.GetLocalClusterDir(clusterName)
-	pluginPath := filepath.Join(rootDir, "node1", "plugins", vmID.String())
-	if err := utils.SetupExecFile(app.Log, vmBin, pluginPath); err != nil {
-		return err
-	}
-
-	cli, err := binutils.NewGRPCClientWithEndpoint(
-		binutils.LocalClusterGRPCServerEndpoint,
-		binutils.WithAvoidRPCVersionCheck(true),
-		binutils.WithDialTimeout(constants.FastGRPCDialTimeout),
-	)
-	if err != nil {
-		return err
-	}
-	ctx, cancel := network.BootstrappingContext()
-	defer cancel()
-	status, err := cli.Status(ctx)
-	if err != nil {
-		return err
-	}
-	for _, nodeInfo := range status.ClusterInfo.NodeInfos {
-		ux.Logger.PrintToUser("Restarting node %s to track newly deployed network", nodeInfo.Name)
-		if err := LocalNodeTrackSubnet(
-			ctx,
-			cli,
-			app,
-			rootDir,
-			avalancheGoBinaryPath,
-			blockchainName,
-			blockchainID,
-			subnetID,
-			nodeInfo.Name,
-		); err != nil {
-			return err
-		}
-		if err := AddNodeInfoToSidecar(&sc, nodeInfo, network); err != nil {
-			return fmt.Errorf("failed to update sidecar with new node info: %w", err)
-		}
-	}
-	ux.Logger.PrintToUser("Waiting for blockchain %s ID %s to be bootstrapped", blockchainName, blockchainID.String())
-	if err := WaitBootstrapped(ctx, cli, blockchainID.String()); err != nil {
-		return fmt.Errorf("failure waiting for local cluster %s bootstrapping: %w", blockchainName, err)
-	}
-	if err := app.UpdateSidecar(&sc); err != nil {
-		return err
-	}
-	ux.Logger.GreenCheckmarkToUser("%s successfully tracking %s", clusterName, blockchainName)
-	return nil
-}
-
 func LocalNodeTrackSubnet(
 	ctx context.Context,
 	cli client.Client,
